Bind dependencies to Hilt's graph
ServiceLocator was previously used to construct the TasksRepository and all
it's needed dependencies and then inject this into ViewModelFactory.

We'll expose this TaskRepository in Hilt's graph as a singleton, with providers
also for all it's dependencies.
Since we're already use Hilt's mechanisms for ViewModel injection all
ViewModels will automagically be injected with the repository.

The ServiceLocator file is to be deleted when it will not be used anymore.